juliet snippets

+++ additional/run_scan-build_forJuliet.pl
@@ -5,7 +5,8 @@ use 5.014;
 use warnings;
 use Time::HiRes;
 
-my $workDir ="/Users/ugurmeryem/Juliet/testcases"; #$ARGV[0];
+my $workDir ="/home/ukoc/juliet/testcases"; #$ARGV[0];
+my $logDir="/home/ukoc/juliet-logs/";
 my @CWEFolders = ();
 
 opendir(DIR, $workDir) or die $!;
@@ -15,19 +16,22 @@ while (my $folder = readdir(DIR)) {
 }
 
 foreach my $folder (@CWEFolders){
+   my $logfile=$logDir.substr($folder, 0, 6)."-scanbuild.log"; 
    if (-d "$workDir/$folder/s01") {
-      next;
       opendir(DIR, "$workDir/$folder") or die $!;
       while (my $subfolder = readdir(DIR)) {
          if($subfolder eq ".." or $subfolder eq "."){ next; }
-         `cd $workDir/$folder/$subfolder && scan-build gcc -c -I ../../../testcasesupport -D OMITBAD *1.c > 000-scan-build.log 2>&1`;
-         `./run_creduce_forJuliet.pl $workDir/$folder 000-scan-build.log`;
-         `cd $workDir/$folder/$subfolder && rm -fr *.o *.orig`;
+	 if (-e "$workDir/$folder/$subfolder/Makefile"){
+	     $logfile=$logDir.substr($folder, 0, 6).$subfolder."-scanbuild.log";
+	     `cd $workDir/$folder/$subfolder && scan-build gcc -c -I ~/juliet/testcasesupport -D OMITBAD *.c > $logfile 2>&1`;
+	     `./run_creduce_forJuliet.pl $workDir/$folder/$subfolder $logfile`;
+	     `cd $workDir/$folder/$subfolder && rm -fr *.o *.orig`;
+	 }
       }
    } elsif (-e "$workDir/$folder/Makefile"){
-      `cd $workDir/$folder && scan-build gcc -c -I ../../testcasesupport -D OMITBAD *1.c > 000-scan-build.log 2>&1`;
-      `./run_creduce_forJuliet.pl $workDir/$folder 000-scan-build.log`;
+       next;
+      `cd $workDir/$folder && scan-build gcc -c -I ~/juliet/testcasesupport -D OMITBAD *.c > $logfile 2>&1`;
+      `./run_creduce_forJuliet.pl $workDir/$folder $logfile`;
       `cd $workDir/$folder && rm -fr *.o *.orig`;
-      exit;
    }
 }
